ref: move handshake protocol skeleton into own module
